Move MeanDeviationIndicator and Test to statistics package
